eth/protocols/snap: extract syncV1 main loop into sync_v1.go
